backend/requirements.txt
Django==4.2.7
django-ninja==1.0.1
django-cors-headers==4.4.0
djongo==1.3.6
pymongo==4.5.0
python-dotenv==1.0.0
Pillow==10.0.1
uvicorn==0.24.0
PyJWT==2.9.0